Begin utilizing SourceCred
This is a PR because it'll break the deploy until a fix is merged in
sourcecred/sourcecred. Once the fix is published, we'll be good to go.

Specifically, the `module` path specified in sourcecred/sourcecred
fails to build with netlify's transpiler. If you want to continue
working with the sourcecred import, just delete the line with
`  "module": "src/api/index.js",`
and netlify will be able to build and deploy the function locally, for
further development